sam_app/requirements.txt
requests
python-telegram-bot==13.11
requests-html
aws-lambda-powertools